Move VisibilityScope into its own file
